requirements.txt
Flask
pymongo[srv]
python-dotenv
Flask-WTF
gunicorn